+++ draft-packages/modal/KaizenDraft/Modal/Primitives/GenericModal.tsx
@@ -1,14 +1,18 @@
 import * as React from "react"
 import { createPortal } from "react-dom"
 import FocusLock from "react-focus-lock"
+import uuid from "uuid/v4"
 const { CSSTransition } = require("react-transition-group")
+import {
+  ModalAccessibleContext,
+  ModalAccessibleContextType,
+} from "./ModalAccessibleContext"
 
 import { warn } from "@kaizen/component-library/util/console"
-import { ID_DESCRIBEDBY, ID_LABELLEDBY } from "./constants"
 
 const styles = require("./GenericModal.scss")
 
-interface Props {
+interface GenericModalContainerProps {
   readonly isOpen: boolean
   readonly children: React.ReactNode
   readonly focusLockDisabled?: boolean
@@ -16,9 +20,32 @@ interface Props {
   readonly onOutsideModalClick?: (event: React.MouseEvent) => void
 }
 
+interface GenericModalProps
+  extends GenericModalContainerProps,
+    ModalAccessibleContextType {}
+
 const MODAL_TRANSITION_TIMEOUT = 350
 
-class GenericModal extends React.Component<Props> {
+function GenericModalContainer(props: GenericModalContainerProps) {
+  const labelledByID = uuid()
+  const describedByID = uuid()
+  return (
+    <ModalAccessibleContext.Provider
+      value={{
+        labelledByID,
+        describedByID,
+      }}
+    >
+      <GenericModal
+        {...props}
+        labelledByID={labelledByID}
+        describedByID={describedByID}
+      />
+    </ModalAccessibleContext.Provider>
+  )
+}
+
+class GenericModal extends React.Component<GenericModalProps> {
   scrollLayer: HTMLDivElement | null = null
   modalLayer: HTMLDivElement | null = null
 
@@ -26,7 +53,7 @@ class GenericModal extends React.Component<Props> {
     if (this.props.isOpen) this.onOpen()
   }
 
-  componentDidUpdate(prevProps: Props) {
+  componentDidUpdate(prevProps: GenericModalProps) {
     const hasJustOpened = !prevProps.isOpen && this.props.isOpen
     const hasJustClosed = prevProps.isOpen && !this.props.isOpen
     if (hasJustOpened) this.onOpen()
@@ -42,6 +69,7 @@ class GenericModal extends React.Component<Props> {
     this.preventBodyScroll()
     this.ensureAccessiblityIsMet()
     this.scrollModalToTop()
+    this.focusAccessibleLabel()
     if (this.modalLayer) {
       this.removeAriaHider = createAriaHider(this.modalLayer)
     }
@@ -99,7 +127,7 @@ class GenericModal extends React.Component<Props> {
     if (!this.modalLayer) return
     // Ensure that consumers have provided an element that labels the modal
     // to meet ARIA accessibility guidelines.
-    if (!document.getElementById(ID_LABELLEDBY)) {
+    if (!document.getElementById(this.props.labelledByID)) {
       warn(
         `When using the Modal component, you must provide a label for the modal.
         Make sure you have a <ModalAccessibleLabel /> component with content that labels the modal.`
@@ -117,6 +145,17 @@ class GenericModal extends React.Component<Props> {
     })
   }
 
+  focusAccessibleLabel() {
+    if (this.modalLayer) {
+      const labelElement: HTMLElement | null = document.getElementById(
+        this.props.labelledByID
+      )
+      if (labelElement) {
+        labelElement.focus()
+      }
+    }
+  }
+
   render(): React.ReactPortal {
     const { isOpen, children, focusLockDisabled = false } = this.props
 
@@ -130,7 +169,11 @@ class GenericModal extends React.Component<Props> {
       >
         {/* This is not an unused div. It will receive `animating-` classes from react-transition-group */}
         <div>
-          <FocusLock disabled={focusLockDisabled} returnFocus={true}>
+          <FocusLock
+            disabled={focusLockDisabled}
+            returnFocus={true}
+            autoFocus={false}
+          >
             <div className={styles.backdropLayer} />
             <div
               className={styles.scrollLayer}
@@ -141,8 +184,8 @@ class GenericModal extends React.Component<Props> {
             >
               <div
                 role="dialog"
-                aria-labelledby={ID_LABELLEDBY}
-                aria-describedby={ID_DESCRIBEDBY}
+                aria-labelledby={this.props.labelledByID}
+                aria-describedby={this.props.describedByID}
                 className={styles.modalLayer}
                 ref={modalLayer => (this.modalLayer = modalLayer)}
               >
@@ -232,4 +275,4 @@ function createAriaHider(dialogNode: HTMLElement): () => void {
   }
 }
 
-export default GenericModal
+export default GenericModalContainer