+++ draft-packages/stories/Tabs.stories.tsx
@@ -1,11 +1,19 @@
 import classnames from "classnames"
 import * as React from "react"
 
-import { Text } from "@kaizen/component-library"
+import { Box, Heading, Paragraph } from "@kaizen/component-library"
 import { Tabs } from "@kaizen/draft-tabs"
 
 export default {
   title: "Tabs (React)",
+  component: Tabs,
+  parameters: {
+    info: {
+      text: `
+      import { Tabs } from "@kaizen/draft-tabs"
+      `,
+    },
+  },
 }
 
 export const Default = () => {
@@ -83,16 +91,23 @@ export const WithCustomTabRenderer = () => {
   ]
   return (
     <div>
-      <Text tag="h1">Custom tab renderer</Text>
-      <Text tag="p">
-        This would most commonly be used when you need to render, for example, a
-        react-router <code>&lt;Link /&gt;</code> instead of the default anchor
-        tag (<code>&lt;a /&gt;</code>) or a custom attribute for your elements.
-      </Text>
-      <Text tag="p">
-        This storybook example renders anchor tags, but you can simply replace
-        it with the correct element for your use case.
-      </Text>
+      <Box mb={1}>
+        <Heading variant="heading-1" tag="h1">
+          Custom tab renderer
+        </Heading>
+      </Box>
+      <Box mb={1}>
+        <Paragraph tag="p" variant="body">
+          This would most commonly be used when you need to render, for example,
+          a react-router <code>&lt;Link /&gt;</code> instead of the default
+          anchor tag (<code>&lt;a /&gt;</code>) or a custom attribute for your
+          elements.
+        </Paragraph>
+        <Paragraph tag="p" variant="body">
+          This storybook example renders anchor tags, but you can simply replace
+          it with the correct element for your use case.
+        </Paragraph>
+      </Box>
       <div>
         <Tabs
           tabs={tabs}