+++ packages/component-library/stories/GlobalNotification.stories.tsx
@@ -4,11 +4,19 @@ import { GlobalNotification } from "@kaizen/component-library"
 
 export default {
   title: "GlobalNotification (React)",
+  component: GlobalNotification,
+  parameters: {
+    info: {
+      text: `
+        import { GlobalNotification } from "@kaizen/component-library";
+      `,
+    },
+  },
 }
 
 export const PositiveKaizenSiteDemo = () => (
   <GlobalNotification type="affirmative" automationId="notification1">
-    New user data, imported by mackenzie@hooli.com has successfully uploaded.
+    New user data, imported by mackenzie@hooli.com has successfully uploaded.{" "}
     <a href="/">Manage users is now available</a>
   </GlobalNotification>
 )
@@ -20,22 +28,20 @@ PositiveKaizenSiteDemo.story = {
 export const Informative = () => (
   <GlobalNotification type="informative" automationId="notification1">
     New user data is currently being processed. We'll let you know when the
-    process is completed.
-    <a href="/">Manage users</a>
+    process is completed. <a href="/">Manage users</a>
   </GlobalNotification>
 )
 
 export const Cautionary = () => (
   <GlobalNotification type="cautionary" automationId="notification1">
     New user data, imported by mackenzie@hooli.com has uploaded with some minor
-    issues.
-    <a href="/">View issues</a>
+    issues. <a href="/">View issues</a>
   </GlobalNotification>
 )
 
 export const Negative = () => (
   <GlobalNotification type="negative" automationId="notification1">
-    No network connection. Check your connection and try again.
+    No network connection. Check your connection and try again.{" "}
     <a href="/">Refresh</a>.
   </GlobalNotification>
 )
@@ -47,21 +53,19 @@ export const MultipleNotifications = () => (
     }}
   >
     <GlobalNotification type="affirmative" automationId="notification1">
-      New user data, imported by mackenzie@hooli.com has successfully uploaded.
+      New user data, imported by mackenzie@hooli.com has successfully uploaded.{" "}
       <a href="/">Manage users is now available</a>
     </GlobalNotification>
     <GlobalNotification type="informative" automationId="notification2">
       New user data is currently being processed. We'll let you know when the
-      process is completed.
-      <a href="/">Manage users</a>
+      process is completed. <a href="/">Manage users</a>
     </GlobalNotification>
     <GlobalNotification type="cautionary" automationId="notification3">
       New user data, imported by mackenzie@hooli.com has uploaded with some
-      minor issues.
-      <a href="/">View issues</a>
+      minor issues. <a href="/">View issues</a>
     </GlobalNotification>
     <GlobalNotification type="negative" automationId="notification4">
-      No network connection. Check your connection and try again.
+      No network connection. Check your connection and try again.{" "}
       <a href="/">Refresh</a>.
     </GlobalNotification>
   </div>