+++ draft-packages/dropdown-menu/CHANGELOG.md
@@ -3,6 +3,108 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.5.2...@kaizen/draft-dropdown-menu@1.5.3) (2020-06-12)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.5.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.5.1...@kaizen/draft-dropdown-menu@1.5.2) (2020-06-11)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.5.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.5.0...@kaizen/draft-dropdown-menu@1.5.1) (2020-06-11)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+# [1.5.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.4.1...@kaizen/draft-dropdown-menu@1.5.0) (2020-06-11)
+
+
+### Features
+
+* replace remaining usages of ca-type-ideal-sans-xxx ([#538](https://github.com/cultureamp/kaizen-design-system/issues/538)) ([4f6618f](https://github.com/cultureamp/kaizen-design-system/commit/4f6618feb2bfe894c56ee994c15d3fddd87c048e))
+
+
+
+
+
+## [1.4.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.4.0...@kaizen/draft-dropdown-menu@1.4.1) (2020-06-10)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+# [1.4.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.3.13...@kaizen/draft-dropdown-menu@1.4.0) (2020-06-09)
+
+
+### Features
+
+* Rollout inter ([#530](https://github.com/cultureamp/kaizen-design-system/issues/530)) ([a1b2059](https://github.com/cultureamp/kaizen-design-system/commit/a1b2059980ea753036caa5cb15ba6b1235d52ba4))
+
+
+
+
+
+## [1.3.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.3.12...@kaizen/draft-dropdown-menu@1.3.13) (2020-06-05)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.3.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.3.11...@kaizen/draft-dropdown-menu@1.3.12) (2020-06-05)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.3.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.3.10...@kaizen/draft-dropdown-menu@1.3.11) (2020-06-05)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.3.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.3.9...@kaizen/draft-dropdown-menu@1.3.10) (2020-06-04)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.3.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.3.8...@kaizen/draft-dropdown-menu@1.3.9) (2020-06-02)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.3.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.3.7...@kaizen/draft-dropdown-menu@1.3.8) (2020-06-02)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
 ## [1.3.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.3.6...@kaizen/draft-dropdown-menu@1.3.7) (2020-05-28)
 
 **Note:** Version bump only for package @kaizen/draft-dropdown-menu