+++ draft-packages/well/CHANGELOG.md
@@ -3,6 +3,102 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.22...@kaizen/draft-well@2.5.23) (2020-06-12)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.22](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.21...@kaizen/draft-well@2.5.22) (2020-06-11)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.20...@kaizen/draft-well@2.5.21) (2020-06-11)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.19...@kaizen/draft-well@2.5.20) (2020-06-11)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.18...@kaizen/draft-well@2.5.19) (2020-06-10)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.17...@kaizen/draft-well@2.5.18) (2020-06-09)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.16...@kaizen/draft-well@2.5.17) (2020-06-05)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.16](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.15...@kaizen/draft-well@2.5.16) (2020-06-05)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.15](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.14...@kaizen/draft-well@2.5.15) (2020-06-05)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.14](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.13...@kaizen/draft-well@2.5.14) (2020-06-04)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.12...@kaizen/draft-well@2.5.13) (2020-06-02)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.5.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.11...@kaizen/draft-well@2.5.12) (2020-06-02)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
 ## [2.5.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.5.10...@kaizen/draft-well@2.5.11) (2020-05-28)
 
 **Note:** Version bump only for package @kaizen/draft-well